+++ tests/computed/flat.isPristine.ts
@@ -11,7 +11,7 @@ export default ($) => {
     it('$G isPristine should be true', () => expect($.$G.isPristine).to.be.true);
     it('$H isPristine should be true', () => expect($.$H.isPristine).to.be.true);
     it('$I isPristine should be true', () => expect($.$I.isPristine).to.be.true);
-    it('$L isPristine should be false', () => expect($.$L.isPristine).to.be.false);
+    it('$L isPristine should be true', () => expect($.$L.isPristine).to.be.true); //
     it('$M isPristine should be false', () => expect($.$M.isPristine).to.be.false);
     it('$N isPristine should be true', () => expect($.$N.isPristine).to.be.true);
   });