+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 5.5.0 (next)
+
+- Updated add()/del()/update() actions to handle `changed` field prop.
+(not triggering anymore `onChange` when using add(), use `onAdd`/`onDel` hooks instead).
+- Updated `changed` computed prop behavior for nested fields and Event Hooks triggering.
+- Events Hooks now are triggered also from actions if not used Event Handlers.
+
+fix: #585 #531
+
 # 5.4.1 (next)
 
 fix: #371 #399 #408